+++ src/blueprint/agent_generator/generator/part_generators/__init__.py
@@ -1,5 +1,4 @@
 from .main_part_generator import MainPartGenerator
-from .agent_map_part_generator import AgentMapPartGenerator
 from .api_part_generator import APIPartGenerator
 from .service_part_generator import ServicePartGenerator
 from .init_part_generator import InitPartGenerator
@@ -13,7 +12,6 @@ from .tests_part_generator import TestsPartGenerator
 
 __all__ = [
     "MainPartGenerator",
-    "AgentMapPartGenerator",
     "APIPartGenerator",
     "ServicePartGenerator",
     "InitPartGenerator",